+++ tests/integration/lookups/v2/test_phone_number.py
@@ -42,6 +42,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
             }
             '''
@@ -70,6 +72,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+141599299600"
             }
             '''
@@ -100,6 +104,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
             }
             '''
@@ -136,6 +142,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -172,6 +180,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -208,6 +218,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -240,6 +252,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -282,6 +296,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "live_activity": null,
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -320,6 +336,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 },
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -362,6 +380,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 },
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -405,6 +425,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 },
                 "line_type_intelligence": null,
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+447772000001"
             }
             '''
@@ -437,6 +459,8 @@ class PhoneNumberTestCase(IntegrationTestCase):
                     "type": "nonFixedVoip"
                 },
                 "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
             }
             '''
@@ -475,6 +499,93 @@ class PhoneNumberTestCase(IntegrationTestCase):
                     "summary_score": 90,
                     "error_code": null,
                     "error_message": null
+                },
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": null,
+                "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
+            }
+            '''
+        ))
+
+        actual = self.client.lookups.v2.phone_numbers("phone_number").fetch()
+
+        self.assertIsNotNone(actual)
+
+    def test_fetch_sms_pumping_risk_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "calling_country_code": "1",
+                "country_code": "US",
+                "phone_number": "+14159929960",
+                "national_format": "(415) 992-9960",
+                "valid": true,
+                "validation_errors": [],
+                "caller_name": null,
+                "sim_swap": null,
+                "call_forwarding": null,
+                "live_activity": null,
+                "line_type_intelligence": null,
+                "identity_match": null,
+                "sms_pumping_risk": {
+                    "carrier": {
+                        "name": "Twilio Carrier",
+                        "mobile_country_code": "214",
+                        "mobile_network_code": "07",
+                        "carrier_risk_score": 69,
+                        "carrier_risk_category": "moderate"
+                    },
+                    "number_blocked": false,
+                    "number_blocked_date": null,
+                    "number_blocked_last_3_months": null,
+                    "error_code": null
+                },
+                "disposable_phone_number_risk": null,
+                "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
+            }
+            '''
+        ))
+
+        actual = self.client.lookups.v2.phone_numbers("phone_number").fetch()
+
+        self.assertIsNotNone(actual)
+
+    def test_fetch_disposable_phone_number_risk_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "calling_country_code": "1",
+                "country_code": "US",
+                "phone_number": "+14159929960",
+                "national_format": "(415) 992-9960",
+                "valid": true,
+                "validation_errors": [],
+                "caller_name": null,
+                "sim_swap": null,
+                "call_forwarding": null,
+                "live_activity": null,
+                "line_type_intelligence": null,
+                "identity_match": null,
+                "sms_pumping_risk": null,
+                "disposable_phone_number_risk": {
+                    "carrier": {
+                        "mobile_country_code": "452",
+                        "mobile_network_code": "05",
+                        "carrier_risk_score": 17,
+                        "carrier_risk_category": "low",
+                        "line_type": "mobile",
+                        "name": "Twilio Carrier"
+                    },
+                    "is_disposable": false,
+                    "disposable_phone_risk_score": 0,
+                    "disposable_phone_risk_category": "low",
+                    "velocity_risk_score": 0,
+                    "velocity_risk_category": "low",
+                    "summary_risk_score": 0,
+                    "summary_risk_category": "low",
+                    "error_code": null
                 },
                 "url": "https://lookups.twilio.com/v2/PhoneNumbers/+14159929960"
             }